Fix connection to TPJ SQL Server DB
The following command now works:

    ./djadmin.sh dbshell --database tpj

The odbc configuration was taken from https://askubuntu.com/a/208383

The current implementation not ideal because the copying of files requires `sync-docker-filesfiles.sh`. Since unixodbc installs empty copies of odbc.ini and odbcinst.ini (and freetds may install freetds.conf) the order of operations is a bit tricky. An initial attempt to use the `volumes` in `docker-compose.yml` failed, possibly due to the order of operations.